requirements.txt
langgraph
langchain
langchain_community
langchainhub
langchain_groq
langchain_huggingface
flask
flask-cors
pandas
numpy
openpyxl
PyPDF2
python-docx
sqlalchemy
pyodbc
lxml
chromadb
sentence-transformers
openai
plotly
seaborn
scikit-learn
python-dotenv
gunicorn
bs4
tiktoken